+++ packages/rollup-plugin/src/index.ts
@@ -6,22 +6,30 @@ import {
   IdentifierOption,
   getSourceFromVirtualCssFile,
   virtualCssFileFilter,
+  CompileOptions,
 } from '@vanilla-extract/integration';
-import { relative, normalize, dirname } from 'path';
+import { posix } from 'path';
+
+const { relative, normalize, dirname } = posix;
 
 interface Options {
   identifiers?: IdentifierOption;
   cwd?: string;
+  esbuildOptions?: CompileOptions['esbuildOptions'];
 }
 export function vanillaExtractPlugin({
   identifiers,
   cwd = process.cwd(),
+  esbuildOptions,
 }: Options = {}): Plugin {
   const emittedFiles = new Map<string, string>();
   const isProduction = process.env.NODE_ENV === 'production';
 
   return {
     name: 'vanilla-extract',
+    buildStart() {
+      emittedFiles.clear();
+    },
     async transform(_code, id) {
       if (!cssFileFilter.test(id)) {
         return null;
@@ -33,17 +41,22 @@ export function vanillaExtractPlugin({
       const { source, watchFiles } = await compile({
         filePath,
         cwd,
+        esbuildOptions,
       });
 
       for (const file of watchFiles) {
         this.addWatchFile(file);
       }
 
-      return processVanillaFile({
+      const output = await processVanillaFile({
         source,
         filePath,
         identOption: identifiers ?? (isProduction ? 'short' : 'debug'),
       });
+      return {
+        code: output,
+        map: { mappings: '' },
+      };
     },
     async resolveId(id) {
       if (!virtualCssFileFilter.test(id)) {
@@ -78,7 +91,7 @@ export function vanillaExtractPlugin({
 
       // ...replace import paths with relative paths to emitted css files
       const chunkPath = dirname(chunkInfo.fileName);
-      return importsToReplace.reduce((codeResult, importPath) => {
+      const output = importsToReplace.reduce((codeResult, importPath) => {
         const assetId = emittedFiles.get(importPath)!;
         const assetName = this.getFileName(assetId);
         const fixedImportPath = `./${normalize(
@@ -86,6 +99,10 @@ export function vanillaExtractPlugin({
         )}`;
         return codeResult.replace(importPath, fixedImportPath);
       }, code);
+      return {
+        code: output,
+        map: chunkInfo.map ?? null,
+      };
     },
   };
 }