+++ packages/rollup-plugin/test/rollup-plugin.test.ts
@@ -5,7 +5,7 @@ import path from 'path';
 
 import { vanillaExtractPlugin } from '..';
 
-async function buildAndMatchSnapshot(outputOptions: OutputOptions) {
+async function build(outputOptions: OutputOptions) {
   const bundle = await rollup({
     input: require.resolve('@fixtures/themed'),
     plugins: [
@@ -19,15 +19,17 @@ async function buildAndMatchSnapshot(outputOptions: OutputOptions) {
   });
   const { output } = await bundle.generate(outputOptions);
   output.sort((a, b) => a.fileName.localeCompare(b.fileName));
+  return output;
+}
+
+async function buildAndMatchSnapshot(outputOptions: OutputOptions) {
+  const output = await build(outputOptions);
   expect(
     output.map((chunkOrAsset) => [
       chunkOrAsset.fileName,
       chunkOrAsset.type === 'asset' ? chunkOrAsset.source : chunkOrAsset.code,
     ]),
   ).toMatchSnapshot();
-  if (bundle) {
-    await bundle.close();
-  }
 }
 
 describe('rollup-plugin', () => {
@@ -51,9 +53,24 @@ describe('rollup-plugin', () => {
     await buildAndMatchSnapshot({
       format: 'esm',
       preserveModules: true,
+      preserveModulesRoot: path.dirname(require.resolve('@fixtures/themed')),
       assetFileNames({ name }) {
-        return name?.replace(/^@fixtures\/themed\/src\//, '') ?? '';
+        return name?.replace(/^src\//, '') ?? '';
       },
     });
+  });
+
+  it('should build with sourcemaps', async () => {
+    const output = await build({
+      format: 'esm',
+      preserveModules: true,
+      sourcemap: true,
+    });
+    expect(
+      output.map((chunkOrAsset) => [
+        chunkOrAsset.fileName,
+        chunkOrAsset.type === 'asset' ? '' : chunkOrAsset.map?.mappings,
+      ]),
+    ).toMatchSnapshot();
   });
 });